Move check and ask for user location services to location utils
